LPD-34441 - Refactor. Final folder rename
Job was done in parallel, with filters and filters_new folder while doing the refactor.
Follow commits to track the progress. This final commit replaces the original 'filters' folder
with the 'filters_new'. Main changes are related to folder names and methods to match the UI.
